model/api/requirements.txt
gradio
numpy
tensorflow
keras
fastapi
requests
Pillow
uvicorn
io